+++ examples/complete/variables.tf
@@ -1,0 +1,13 @@
+variable "tags" {
+  type        = map(string)
+  description = "Tags to apply to the resources"
+  default = {
+    Environment        = "example"
+    "user::CostCenter" = "terraform"
+    Department         = "DevOps"
+    Project            = "Examples"
+    Owner              = "Boldlink"
+    LayerName          = "cExample"
+    LayerId            = "cExample"
+  }
+}